pyproject.toml
[project]
name = "bigbag"
version = "0.1.0"
description = "BigBag"
authors = []
dependencies = [
    "flask>=3.1.2",
    "mysql-connector-python>=9.4.0",
    "flask-login>=0.6.3",
]
requires-python = ">=3.13"
readme = "README.md"

[tool.ruff]
line-length = 79
exclude = ["tests/*"]

[tool.ruff.lint.pycodestyle]
max-line-length = 79

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["E402", "F401"]
"**/{tests,docs,tools}/*" = ["E402"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 72

[tool.pytest.ini_options]
addopts = [
  "--cov-report=term-missing",
  "--cov-report=html",
  "--cov-fail-under=100",
]
